Move the /bags tests into a separate file
